requirements.txt
seleniumbase
pyautogui
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sportsreference
termcolor
